refactor(browser): split Playwright session manager
